telegram_gpt3_bot.py
import os
import openai
import telegram
from telegram.ext import Updater, MessageHandler, CommandHandler, Filters

#set OpenAI API key
openai.api_key = "your_openai_api_key_here"

#initialize Telegram bot
bot = telegram.Bot(token="your_telegram_api_key_here")

def generate_response(prompt):
  completions = openai.Completion.create(
    engine="text-davinci-002",
    prompt=prompt,
    max_tokens=1024,
    n=1,
    stop=None,
    temperature=0.5,
  )

  message = completions.choices[0].text
  return message

#define a function that handles incoming messages from Telegram
def handle_message(update, context):
  text = update.message.text
  response = generate_response(text)
  chat_id = update.message.chat_id
  bot.send_message(chat_id=chat_id, text=response)

#set up the Telegram bot to listen for incoming messages
updater = Updater("<YOUR_BOT_TOKEN>", use_context=True)
dispatcher = updater.dispatcher
message_handler = MessageHandler(Filters.text, handle_message)
dispatcher.add_handler(message_handler)

#start the bot
updater.start_polling()