requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
APScheduler==3.6.3
async-timeout==4.0.2
attrs==22.2.0
cachetools==4.2.2
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
install==1.3.5
multidict==6.0.4
openai==0.26.4
pyTelegramBotAPI==4.9.0
python-dotenv==0.21.1
python-telegram-bot==13.7
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tornado==6.2
tqdm==4.64.1
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
yarl==1.8.2
